aiops-seed-code/bedrock-agents/model_build/ml_pipelines/requirements.txt
# Core dependencies
boto3>=1.34.0
sagemaker>=2.200.0
pyyaml>=6.0

# AWS SDK for Bedrock
# Note: boto3 includes bedrock and bedrock-agent services

# Utilities
requests>=2.31.0
jsonschema>=4.20.0
